requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
python-dotenv>=1.0.1
qdrant-client>=1.16.0
openai>=2.20.0
pydantic>=2.12.0

